CORE-2063 Force sign-out if Keycloak access token is expired
The `next-auth` session and jwt `exp` values are 30 days longer than
the Keycloak access token's `exp` value,
so after the access token expires,
the session has not expired and the user still appears to be logged in,
but calling a terrain endpoint with that token will return a 403 error.

So this adds another `accessTokenExp` field to the `session` object
that can be checked in a `useEffect` in the `AccountAvatar`,
which can call the `signOut` function if the token is expired.