CORE-2094 Update `/api/authznet/callbacks` to update subscription
Updated `/api/authznet/callbacks` to update subscription and addons on
successful payment, and to email the user and admins.

Refactored the `serviceAccountEmailReceipt` function in
`@/app/api/terrain` into separate functions.

Updated the email functions to use `OrderDetails` fetched from the db
instead of the `OrderRequest` submitted to Authorize.net.

Updated `serviceAccountEmailReceipt` to lookup the user's email with
the username, since the user session is not available in the callback.

Added `serviceAccountFetchUserInfo` and
`serviceAccountFetchSubscription` to `@/app/api/terrain`.

Refactored the `getPurchaseByPoNumber` in `@/db` to return the same
info as `getUserPurchase`. Both now also return the `id` and `username.