CORE-2094 Update `Checkout` component for hosted forms
Added `SP_AUTHORIZE_NET_HOSTED_ENDPOINT` config and updated the
`Checkout` form to submit the hosted form token to the hosted payment
site with a hidden form input.

Removed the `OrderConfirmation` page.

Removed the `AddressForm` and `PaymentForm` from the `Checkout` form,
and made the `Review` component's content the main content of the
`Checkout` form, which made the `Stepper` and `InfoMobile` obsolete.

Also removed the `dirty` flag from the `BackButton`,
since the `termsAcknowledged` checkbox is the only form element left.